fix(formbuilder): move styling and helper functions to its own files
