+++ src/components/specific/useFormBuilderStyles.ts
@@ -1,0 +1,42 @@
+import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      '& > *': {
+        margin: theme.spacing(1),
+        width: '20ch',
+      },
+    },
+    subcontainer: {
+      flexGrow: 1,
+      maxWidth: 752,
+      margin: theme.spacing(4, 0, 2),
+      backgroundColor: 'rgba(255, 255, 255, 0.10)',
+      borderWidth: '1px',
+      borderStyle: 'solid',
+      boxShadow: '0 0 10px rgba(0, 0, 0, 0.3)',
+      padding: theme.spacing(1),
+      position: 'relative',
+      minHeight: '45px',
+    },
+    input: {
+      '& > *': {
+        width: '500px',
+      },
+    },
+    button: {
+      '& > *': {
+        margin: theme.spacing(1),
+        width: '2ch',
+      },
+    },
+    wrapper: {
+      display: 'grid',
+      gridTemplateColumns: '400px 800px 200px',
+    },
+    column: {
+      padding: theme.spacing(2),
+    },
+  }),
+);
